refac: move userinfo error dialog
Separate  userinfo error dialog from signup form and move to `components/`
