+++ components/Userinfo/ErrorDialog/style.module.scss
@@ -1,0 +1,11 @@
+.wrong {
+  color: $err-color;
+  font-size: large;
+  border-radius: 0.5rem;
+  & button {
+    margin-left: 0.5rem;
+    border-radius: 0.5rem;
+    border: 0.2rem solid $err-color;
+    padding: 0.3rem;
+  }
+}
